making a mono repo, one app is the exercise ready to zip up and the other is the ghpages up with download button to zip file
