+++ .gitignore
@@ -15,6 +15,7 @@ test-results
 
 # production
 /build
+.zip
 
 # misc
 .DS_Store
@@ -31,6 +32,7 @@ yarn-error.log*
 
 # vercel
 .vercel
+
 
 # typescript
 *.tsbuildinfo
