+++ apps/download/index.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html>
+  <head> </head>
+  <body>
+    <h1>React Exercise Download</h1>
+    <br />
+    <a href="/public/assets/react-exercise.zip" download>
+      <button>Download</button>
+    </a>
+  </body>
+</html>
