+++ apps/download/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "react-exercise-download",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "build": "next build && next export"
+  },
+  "dependencies": {
+    "@cruk/cruk-react-components": "^5.0.5",
+    "@emotion/is-prop-valid": "^1.3.0",
+    "@hookform/resolvers": "^3.3.4",
+    "@tanstack/react-query": "^4.3.9",
+    "next": "^14",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-hook-form": "^7.46.1",
+    "styled-components": "^6.1.11",
+    "zod": "^3.22.3"
+  },
+  "devDependencies": {
+    "@axe-core/playwright": "^4.6.0",
+    "@cruk/eslint-config": "^3",
+    "@playwright/test": "^1.39.0",
+    "@types/node": "^20",
+    "@types/react": "18.3.4",
+    "@typescript-eslint/eslint-plugin": "^7",
+    "@typescript-eslint/parser": "^7",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^28.2.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-promise": "^6",
+    "prettier": "^3",
+    "typescript": "^5.0.2"
+  }
+}
