+++ apps/react-exercise/.nvmrc
@@ -1,0 +1,1 @@
+20
