+++ apps/react-exercise/src/contexts/ReactQueryProvider.tsx
@@ -1,0 +1,27 @@
+"use client";
+
+import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
+import { useState } from "react";
+
+export const ReactQueryProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [queryClient] = useState(
+    () =>
+      new QueryClient({
+        defaultOptions: {
+          queries: {
+            refetchOnMount: true,
+          },
+        },
+      }),
+  );
+
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+};
+
+export default ReactQueryProvider;
