+++ apps/react-exercise/src/server/StyledComponentsRegistry.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import React, { useState } from "react";
+import { useServerInsertedHTML } from "next/navigation";
+import { ServerStyleSheet, StyleSheetManager } from "styled-components";
+
+type StyledComponentsRegistryProps = {
+  children: React.ReactNode;
+};
+
+export const StyledComponentsRegistry = ({
+  children,
+}: StyledComponentsRegistryProps) => {
+  const [styledComponentsStyleSheet] = useState(() => new ServerStyleSheet());
+
+  useServerInsertedHTML(() => {
+    const styles = styledComponentsStyleSheet.getStyleElement();
+    styledComponentsStyleSheet.instance.clearTag();
+    return <>{styles}</>;
+  });
+
+  if (typeof window !== "undefined") return <>{children}</>;
+
+  return (
+    <StyleSheetManager
+      sheet={styledComponentsStyleSheet.instance}
+      enableVendorPrefixes
+    >
+      {children}
+    </StyleSheetManager>
+  );
+};
