+++ package.json
@@ -1,42 +1,17 @@
 {
   "name": "react-exercise",
-  "version": "0.1.0",
   "private": true,
+  "description": "Cancer Research UK Online React Exercise",
+  "main": "index.js",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build && next export",
-    "start": "next start",
-    "lint": "next lint",
-    "test": "npx playwright test",
-    "test:debug": "npx playwright test .src/tests --project='chromium' --workers 1 --headed --debug"
+    "dev:react-exercise": "npm run dev -w apps/react-exercise",
+    "build:react-exercise": "npm run build -w apps/react-exercise",
+    "test:react-exercise": "npm run test -w apps/react-exercise",
+    "test:debug:react-exercise": "npm run test:debug -w apps/react-exercise",
+    "zipit": "sh zipup.sh"
   },
-  "dependencies": {
-    "@cruk/cruk-react-components": "^5.0.5",
-    "@emotion/is-prop-valid": "^1.3.0",
-    "@hookform/resolvers": "^3.3.4",
-    "@tanstack/react-query": "^4.3.9",
-    "next": "^14",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-hook-form": "^7.46.1",
-    "styled-components": "^6.1.11",
-    "zod": "^3.22.3"
-  },
-  "devDependencies": {
-    "@axe-core/playwright": "^4.6.0",
-    "@cruk/eslint-config": "^3",
-    "@playwright/test": "^1.39.0",
-    "@types/node": "^20",
-    "@types/react": "18.3.4",
-    "@typescript-eslint/eslint-plugin": "^7",
-    "@typescript-eslint/parser": "^7",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jest": "^28.2.0",
-    "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-promise": "^6",
-    "prettier": "^3",
-    "typescript": "^5.0.2"
-  }
+  "workspaces": [
+    "apps/react-exercise",
+    "apps/react-exercise-download"
+  ]
 }
